# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'lciii-scaffolding'","description":"'LCII Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'LCII Initial Scaffolding'"}],"paths":{"/api/countrie":{"post":{"tags":["country-controller"],"operationId":"createCountry","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AmountOfCountryToSaveDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDto"}}}}}}}},"/api/countries":{"get":{"tags":["country-controller"],"operationId":"getCountriesAllByName","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDto"}}}}}}}},"/api/countries/{language}/language":{"get":{"tags":["country-controller"],"operationId":"Language","parameters":[{"name":"language","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDto"}}}}}}}},"/api/countries/{continent}/continent":{"get":{"tags":["country-controller"],"operationId":"getCountriesDtoContinent","parameters":[{"name":"continent","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryDto"}}}}}}}},"/api/countries/most-borders":{"get":{"tags":["country-controller"],"operationId":"mostBorder","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CountryDto"}}}}}}}},"components":{"schemas":{"AmountOfCountryToSaveDto":{"required":["amountOfCountryToSave"],"type":"object","properties":{"amountOfCountryToSave":{"maximum":10,"minimum":1,"type":"integer","format":"int32"}}},"CountryDto":{"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"}}}}}}